Passage to Spawner Parameter Estimates

The passage to spawner model, documented in vignette("passage_to_spawner_submodel", package = "SRJPEmodel"), estimates the relationship between upstream passage counts and spawner counts (either redd or holding) for several streams. When fit, the model produces an object P2S_model_fits that contains the parameter names, mean parameter estimate, confidence intervals, and diagnostics. Currently, this model is only fit for battle creek, clear creek, deer creek, and mill creek.

P2S_model_fits

Format

A tibble with 343 rows and 12 columns

par_names: Parameter name

mean: Mean of the posterior distribution for a parameter

se_mean: Monte Carlo standard error for summary of all chains merged. See details

sd: Standard deviation of the posterior distribution for a parameter

2.5%: 2.5% quantile of posterior distribution for a parameter

25%: 25% quantile of posterior distribution for a parameter

50%: 50% quantile of posterior distribution for a parameter

75%: 75% quantile of posterior distribution for a parameter

97.5%: 97.5% quantile of posterior distribution for a parameter

n_eff: Effective sample size for a parameter

Rhat: Split Rhats for a parameter

stream: Stream name for the parameter estimate

year_index: Identifier for a parameter estimate (1:n, where n is the total number of years modeled) that allows us to join the observed year back to the data table

year: Year associated with the parameter estimates and observed data