Prepare inputs for pCap STAN model

Prepare inputs for pCap STAN model

prepare_pCap_inputs(
  mainstem = c(FALSE, TRUE),
  mainstem_site = NULL,
  input_catch_data = NULL,
  input_efficiency_data = NULL
)

Arguments

mainstem

Whether or not you want to evaluate efficiency trials for a mainstem site (knights landing, tisdale, and red bluff diversion dam) or for a tributary site. If FALSE, the mark recapture dataset will be filtered to exclude those mainstem sites.

mainstem_site

If you are fitting the model with mainstem == TRUE, you must supply a mainstem site for which to prepare inputs. Can be either knights landing or tisdale.

input_catch_data

Optional argument for weekly catch data. Defaults to SRJPEdata::weekly_juvenile_abundance_catch_data. If passed in, structure of data frame must match that of the default.

input_efficiency_data

Optional argument for weekly efficiency data. Defaults to SRJPEdata::SRJPEdata::weekly_juvenile_abundance_efficiency_data. If passed in, structure of data frame must match that of the default.

Value

a list:

pCap_inputs a list of data and inits for input into the pCap model.

sites_fit a list of site names associated with ind_trib.

Details

This function prepares data for input into a pCap STAN model.

See also