Prepare inputs for survival model

Prepare inputs for survival model

prepare_survival_inputs(number_of_water_year_types = NULL, effect = NULL)

Arguments

number_of_water_year_types

either 2, 3, or leave empty. If empty, the function will call a version of the survival with no covariate effect.

effect

one of no_biological_effect, fork_length_effect, weight_effect, or condition. If supplying an effect, you must supply an argument for number_of_water_year_types

Value

a named list containing the inputs to the run_survival_model() function:

data_inputs a named list of data inputs for passing to the STAN model.

inits a list of initial values for passing to the STAN model.

parameters a vector of parameter names to save from calling the STAN model.

Details

This function prepares the data, initial values, and parameter list needed to call the survival STAN model. It will prepare these data based on what model you want to run.

See also