Call BT-SPAS-X model on all site/run years

Call BT-SPAS-X model on all site/run years

run_multiple_bt_spas_x(
  bt_spas_x_bayes_params,
  bt_spas_x_input_data,
  effort_adjust,
  mainstem_version,
  bugs_directory,
  debug_mode
)

Arguments

bt_spas_x_input_data

data frame containing the same variables as ?SRJPEdata::weekly_juvenile_abundance_model_data

effort_adjust

whether or not you want to use catch adjusted by effort

mainstem_version

whether or not this is run on mainstem tributaries

bugs_directory

where the WinBUGS.exe file can be found. Needs to end in /WinBUGS

debug_mode

whether you want to run bugs in debug mode.

bt_spas_x_bayes_params:

a list containing number_mcmc, number_burnin, number_thin, and number_chains. Can use SRJPEmodel::bt_spas_x_bayes_params.

Value

a list containing the results for every unique site and run year combination in the input data.

Details

This calls run_single_bt_spas_x() on all site and run year combinations present in the input data.